requirements.txt
fastapi
uvicorn
python-dotenv
requests
psycopg2-binary
pydantic
twilio